API Documentation

YardSale.EYSM_base_full — Method

EYSM_base_full(N, W_N, chi, zeta, f, steps,seed; w=nothing,initial_conditions="uniform",
save_every=nothing)

Runs a Monte Carlo simulation of the Extended Yard-Sale Model and returns the whole time series of the wealth distribution.

Arguments

N::Integer: Number of agents.
W_N::Real: Mean wealth of the agents. It is the total wealth divided by the number
of agents.
chi::Real: Taxation rate.
zeta::Real: Wealth-Attained-Advantage parameter.
f::Real: Fraction of the wealth that is redistributed.
steps::Integer: Number of steps of the simulation, measured in Monte Carlo steps.
seed::Integer: Seed for the random number generator.

Optional arguments

w::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
Only used if initial_conditions="custom". Default is nothing.
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the w argument must be provided.
save_every::Union{Nothing, Integer}=nothing: Save the wealth distribution every
save_every steps. Default is nothing, which means saving every N steps.

Returns

Details

This function tries to reproduce as closely as possible the results of the original EYSM proposed by Boghosian et al. in 2017. The model has two parameters, and it is defined for a fully connected network of $N$ agentes, each one with a wealth $w_i$, such that the total wealth $\sum_i w_i = W$ is conserved.

The interaction rule is defined as follows:

Randomly select two agents i and j.

The amount of wealth exchanged is given by a fraction of the minimum between

the wealths of the two agents.

\[\Delta w = f\min(w_i, w_j)\eta_{ij}\]

where $\eta_{ij}$ is a stochastic variable with values -1 or 1. The expected value of $\eta_{ij}$ is biased towards the richest agent, and it is given by

\[\langle\eta_{ij}\rangle = \zeta \frac{w_i - w_j}{W}\]

The wealth of the agents is updated as follows:

\[w_i \leftarrow w_i - \Delta w + \chi (\frac{W}{N} - w_i) w_j \leftarrow w_j + \Delta w + \chi (\frac{W}{N} - w_j) ```` where ``\chi`` represents the taxation and redistribution rate. In the original paper, the authors work analitically, using a Fokker-Planck approach. In this implementation, we use a Monte Carlo simulation to obtain similar results. # Examples\]

julia wt1 = EYSMbasefull(1000, 1, 0.1, 0.1, 1, 1000, 42) wt2 = EYSMbasefull(64, 1.0f0, 0.1f0, 0.1f0, 1, 1000, 42, initialconditions="random") wt3 = EYSMbasefull(32, 1, 0.1, 0.1, 1, 1000, 42, initialconditions="custom", w=rand(1000)) ```

YardSale.EYSM_base_redistribution — Method

EYSM_base_redistribution(wi, W_N, chif_N)

Calculates the redistribution term in the EYSM model for an agent i with wealth wi.

Arguments

wi::Real: Wealth of agent i.
W_N::Real: Total wealth per agent.
chif_N::Real: Taxation and redistribution rate.

Details

In the EYSM model, the redistribution term of the wealth update of agent i $is$ given by

\[\frac{\chi f}{N}(\frac{W}{N} - w_i)\]

where f is the fraction of the minimum wealth exchanged between agents, and $f$ is playing the role of the Δt term in the Fokker-Planck approach, as the temporal unit of time of the Monte Carlo simulation. $chi$ represents the taxation and redistribution rate and N is the number of agents.

Returns

redist::Real: Redistribution term for agent i.

YardSale.dxdt_net! — Method

dxdt_net!(dxdt, x, p, t)

In-place function to calculate the derivatives of the network model.

Arguments

dxdt::Vector{<:Real}: Vector of derivatives.
x::Vector{<:Real}: Vector of wealths.
p::Tuple{<:Integer, <:Integer, Vector{Vector{<:Integer}},
Vector{<:Integer}, Vector{<:Array}, ::Real, ::Real}: Parameters.
t::Float64: Time.

Details

The network model is given by the following set of ODEs:

\[\dot{x}_i = \frac{T}{N} (- \beta_i x_i + \sum_j^N \beta_j x_j) + \sum_j^N \kappa_{ij} (x_i - x_j) \min(x_i, x_j)\]

where $x_i$ is the wealth of node $i$, $T$ is temperature, $N$ is the number of nodes, $\beta_i$ is the redistribution parameter of node $i$, $\kappa_{ij}$ is the exchange parameter between nodes $i$ and $j$.

YardSale.get_kappa_beta — Method

get_kappa_beta(g::SimpleGraph{<:Integer},interaction_mode::String,taxation_mode::String)

Return the κ matrix and the β vector for the ODEs, given a graph, an Interaction Mode and a Taxation Mode.

Arguments

g::SimpleGraph{<:Integer}: Undirected graph.
interaction_mode::String: Interaction Mode. Can be "A" or "B".
taxation_mode::String: Taxation Mode. Can be "A" or "B".

Returns

Tuple{Vector{Float64},Vector{Float64}}: κ matrix and β vector.

Examples

g = complete_graph(4)
YardSale.get_kappa_beta(g,"A","A")

YardSale.mc_set_initial_conditions — Function

mc_set_initial_conditions(N, W_N, initial_conditions,w=nothing)

Set the initial conditions for the Monte Carlo simulation.

Arguments

N::Integer: Number of agents.
W_N::Real: Total wealth per agent.
initial_conditions::String: Initial condition. Options are "uniform", "random", "noisy"
and "custom".
w::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution. Only used if
initial_conditions="custom". Default is nothing.

Returns

w::Vector{Real}: Initial wealth distribution.

YardSale.rescale_t — Method

rescale_t(t::AbstractVector,N::Integer)

Rescale the ODE time vector to MC steps.

Arguments

t::AbstractVector: Time vector of the ODE solution.
N::Integer: Number of nodes in the network.

Returns

rescaled_t::AbstractVector: Rescaled time vector.

YardSale.solve_ode_net — Method

solve_ode_net(g, tspan, integrator, interaction_mode, taxation_mode, T, seed; kwargs...)

Solves the ODE for the network model.

Arguments

g::SimpleGraph{<:Integer}: Graph.
tspan::Tuple{Float64, Float64}: Tuple with initial and final time.
interaction_mode::String: Interaction mode.
taxation_mode::String: Taxation mode.
T::Real: Temperature.
seed::Integer: Random seed.

Optional arguments

`integrator::SciMLAlgorithm`: An instante of the integrator to use. Default is Tsit5().
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the x0 argument must be provided.
x0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
`kwargs...`: Additional arguments for the solver.

Details

This function solves the ODE for the network model. It calculates the kappa and beta parameters, sets the initial conditions, and solves the ODE. It returns the solution in the standard DifferentialEquations.jl format. A brief description of the ODE approximation model can be found in the dxdt_net! function. The function uses the DifferentialEquations.jl package to solve the ODE. All the parameters for the solver can be passed as keyword arguments.

The initial conditions can be set to "noisy", "random", "uniform", or a custom vector.

"noisy": Initial conditions are set to a random value around 1/N.

$x = (1/N) * (1 + \epsilon)$ where ϵ is white noise with μ=0 and σ=0.01.

"random": Initial conditions are set to a random value, normalized to sum to 1.

"uniform": Initial conditions are set to 1/N.

"custom": Initial conditions are set to a custom vector. The vector must sum to 1.

Also, $x_i$ must be positive for all $i$.

Returns

sol::ODESolution: Solution of the ODE.

Examples

using DifferentialEquations, Graphs, YardSale
# Create a graph
g = erdos_renyi(100, 0.1, seed=42)
interaction_mode, taxation_mode = "A","A"
T = 1.0
seed = 42
tspan = (0.0, 10.0)
sol1 = solve_ode_net(g, tspan, interaction_mode, taxation_mode, T, seed)
sol2 = solve_ode_net(g, tspan, interaction_mode, taxation_mode, T, seed; integrator=RK4())
sol3 = solve_ode_net(g, tspan, interaction_mode, taxation_mode, T, seed;
integrator=RK4(), reltol=1e-6, abstol=1e-6)

YardSale.solve_ode_net_SS — Method

solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed;
initial_conditions="noisy",integrator=Tsit5(), kwargs...)

Solves the ODE for the network model using a steady state solver.

Arguments

g::SimpleGraph{<:Integer}: Graph.
interaction_mode::String: Interaction mode.
taxation_mode::String: Taxation mode.
T::Real: Temperature.
seed::Integer: Random seed.

Optional arguments

`initial_conditions::Union{String, Vector{<:Real}}`="noisy": Initial conditions
for the ODE. Default is "noisy", which sets the initial conditions to a random value
around 1/N. Other options are "random" and "uniform". If a vector is passed, it will be
used as the initial conditions.
`x0::Union{Nothing, Vector{<:Real}}`=nothing: Initial wealth distribution.
`integrator::SciMLAlgorithm`: An instance of the integrator to use. Default is Tsit5().
`kwargs...`: Additional arguments for the solver.

Details

This function solves the ODE for the network model using a steady state solver. It is similar to solve_ode_net, but it uses the DynamicSS solver to find the steady state. Instead of an ODE problem, it uses a SteadyStateProblem to solve the ODE. Reference: https://docs.sciml.ai/DiffEqDocs/stable/types/steadystatetypes/

As in solve_ode_net, the initial conditions can be set to "noisy", "random", "uniform", or a custom vector.

Returns

sol::ODESolution: Solution of the ODE.

Examples

using DifferentialEquations, Graphs, YardSale
# Create a graph
g = erdos_renyi(100, 0.1, seed=42)
interaction_mode, taxation_mode = "A","A"
T = 1.0
seed = 42
sol1 = solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed)
sol2 = solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed; integrator=RK4())
sol3 = solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed;
                            integrator=RK4(),reltol=1e-6, abstol=1e-6
                            )

YardSale.Δw — Method

Δw(f, wi, wj)

Calculate the amount of wealth exchanged between agents i and j in the EYSM model.

Arguments

f::Real: Fraction of the minimum wealth exchanged between agents i and j.
wi::Real: Wealth of agent i.
wj::Real: Wealth of agent j.

Returns

Δw::Real: Amount of wealth exchanged between agents i and j.

YardSale.ηij — Method

ηij(wi, wj, zeta_W)

Calculate the stochastic variable ηij in the EYSM model.

Arguments

wi::Real: Wealth of agent i.
wj::Real: Wealth of agent j.
zeta_W::Real: Normalized bias towards the richest agent.

Details

The expected value of ηij is given by

\[\langle\eta_{ij}\rangle = \zeta \frac{w_i - w_j}{W}\]

This variable is used to determine the direction of the wealth exchange between agents i and j, and can take values ±1.

Returns

ηij::Real Stochastic variable ηij.