Monte Carlo Simulations

YardSale.EYSM_base_callbacks — Method

EYSM_base_callbacks(N, W_N, chi, zeta, f, steps, seed; w=nothing,
initial_conditions="uniform",save_every=nothing, callbacks=nothing,
callbacks_only=false)

Runs a Monte Carlo simulation of the Extended Yard-Sale Model on a fully connected network and returns the whole time series of the wealth distribution and the results of the callbacks.

Arguments

N::Integer: Number of agents.
W_N::Real: Mean wealth of the agents. It is the total wealth divided by the number
of agents.
chi::Real: Taxation rate.
zeta::Real: Wealth-Attained-Advantage parameter.
f::Real: Fraction of the wealth that is redistributed.
steps::Integer: Number of steps of the simulation, measured in Monte Carlo steps.
seed::Integer: Seed for the random number generator.

Optional arguments

w0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
Only used if initial_conditions="custom". Default is nothing.
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the w argument must be provided.
save_every::Union{Nothing, Integer}=nothing: Save the wealth distribution every
save_every steps. Default is nothing, which means saving every N steps.
callbacks::Union{Nothing, Dict{Symbol, Function}}=nothing: Dictionary of callbacks.
The keys are the names of the callbacks, and the values are the functions that will be
called at each checkpoint. The functions must receive the wealth distribution
as an argument.

Returns

w_t::Union{Nothing, Matrix{Real}}: Time series of the wealth distribution. Each row is a checkpoint.
callback_results::Union{Nothing, Dict{Symbol, AbstractVector}}: Results of the callbacks.

YardSale.EYSM_base_full — Method

EYSM_base_full(N, W_N, chi, zeta, f, steps,seed; w=nothing,initial_conditions="uniform",
save_every=nothing)

Runs a Monte Carlo simulation of the Extended Yard-Sale Model and returns the whole time series of the wealth distribution.

Arguments

N::Integer: Number of agents.
W_N::Real: Mean wealth of the agents. It is the total wealth divided by the number
of agents.
chi::Real: Taxation rate.
zeta::Real: Wealth-Attained-Advantage parameter.
f::Real: Fraction of the wealth that is redistributed.
steps::Integer: Number of steps of the simulation, measured in Monte Carlo steps.
seed::Integer: Seed for the random number generator.

Optional arguments

w0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
Only used if initial_conditions="custom". Default is nothing.
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the w0 argument must be provided.
save_every::Union{Nothing, Integer}=nothing: Save the wealth distribution every
save_every steps. Default is nothing, which means saving every N steps.

Returns

Details

This function tries to reproduce as closely as possible the results of the original EYSM proposed by Boghosian et al. in 2017. The model has two parameters, and it is defined for a fully connected network of $N$ agentes, each one with a wealth $w_i$, such that the total wealth $\sum_i w_i = W$ is conserved.

The interaction rule is defined as follows:

Randomly select two agents i and j.

The amount of wealth exchanged is given by a fraction of the minimum between

the wealths of the two agents.

\[\Delta w = f\min(w_i, w_j)\eta_{ij}\]

where $\eta_{ij}$ is a stochastic variable with values -1 or 1. The expected value of $\eta_{ij}$ is biased towards the richest agent, and it is given by

\[\langle\eta_{ij}\rangle = \zeta \frac{w_i - w_j}{W}\]

The wealth of the agents is updated as follows:

\[w_i \to w_i - \Delta w + \chi (\frac{W}{N} - w_i) w_j \to w_j + \Delta w + \chi (\frac{W}{N} - w_j) ```` where ``\chi`` represents the taxation and redistribution rate. In the original paper, the authors work analitically, using a Fokker-Planck approach. In this implementation, we use a Monte Carlo simulation to obtain similar results. # Examples\]

julia wt1 = EYSMbasefull(1000, 1, 0.1, 0.1, 1, 1000, 42) wt2 = EYSMbasefull(64, 1.0f0, 0.1f0, 0.1f0, 1, 1000, 42, initialconditions="random") wt3 = EYSMbasefull(32, 1, 0.1, 0.1, 1, 1000, 42, initialconditions="custom", w=rand(1000)) ```

YardSale.EYSM_net_callbacks — Method

EYSM_net_callbacks(g::SimpleGraph,W_N, interaction_mode, taxation_mode, chi, zeta, f,
steps, seed; w=nothing,initial_conditions="uniform",save_every=nothing,
callbacks=nothing,callbacks_only=false)

Runs a Monte Carlo simulation of the Extender Yard-Sale model on an arbitrary network and returns the whole time series of the wealth distribution and the results of the callbacks.

Arguments

g::SimpleGraph{<:Integer}: Graph representing the network of agents.
W_N::Real: Mean wealth of the agents. It is the total wealth divided by the number
of agents.
interaction_mode::String: Interaction mode. Options are "A" and "B".
taxation_mode::String: Taxation mode. Options are "A" and "B".
chi::Real: Taxation rate.
zeta::Real: Wealth-Attained-Advantage parameter.
f::Real: Fraction of the wealth that is redistributed.
steps::Integer: Number of steps of the simulation, measured in Monte Carlo steps.
seed::Integer: Seed for the random number generator.

Optional arguments

w0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
Only used if initial_conditions="custom". Default is nothing.
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the w argument must be provided.
save_every::Union{Nothing, Integer}=nothing: Save the wealth distribution every
save_every steps. Default is nothing, which means saving every N steps.
callbacks::Union{Nothing, Dict{Symbol, Function}}=nothing: Dictionary of callbacks.
The keys are the names of the callbacks, and the values are the functions that will be
called at each checkpoint. The functions must receive the wealth distribution as an
argument.

Returns

w_t::Union{Nothing, Matrix{Real}}: Time series of the wealth distribution.
Each row is a checkpoint.
callback_results::Union{Nothing, Dict{Symbol, AbstractVector}}: Results of the callbacks.

YardSale.EYSM_net_full — Method

EYSM_net_full(g::SimpleGraph,W_N, interaction_mode, taxation_mode, chi, zeta, f, steps,
seed; w=nothing,initial_conditions="uniform",save_every=nothing)

Runs a Monte Carlo simulation of the Extender Yard-Sale model on an arbitrary network and returns the whole time series of the wealth distribution.

Arguments

g::SimpleGraph{<:Integer}: Graph representing the network of agents.
W_N::Real: Mean wealth of the agents. It is the total wealth divided by the number
of agents.
interaction_mode::String: Interaction mode. Options are "A" and "B".
taxation_mode::String: Taxation mode. Options are "A" and "B".
chi::Real: Taxation rate.
zeta::Real: Wealth-Attained-Advantage parameter.
f::Real: Fraction of the wealth that is redistributed.
steps::Integer: Number of steps of the simulation, measured in Monte Carlo steps.
seed::Integer: Seed for the random number generator.

Optional arguments

w0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
Only used if initial_conditions="custom". Default is nothing.
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the w argument must be provided.
save_every::Union{Nothing, Integer}=nothing: Save the wealth distribution every
save_every steps. Default is nothing, which means saving every N steps.

Details

This function is the generalization of the EYSMbasefull function to arbitrary networks. In the extension, an ambiguity arises in the definition of the interaction rule and the taxation rule. That is why we have to define previously the way in which agents will be selected to exchange wealth and the way in which the taxation will be applied. For each of these two steps, we have two options, which are defined by the interactionmode and taxationmode. The interaction_mode defines the way in which agents will be selected to exchange wealth. The options are:

"A": Randomly select a link.

"B": Randomly select a node and one of its neighbors.

The taxation_mode defines the way in which the taxation will be applied. The options are:

"A": Tax exchanging agents.

"B": Tax two random agents.

The equation governing the wealth exchange is a generalization of the one used in the EYSMbasefull function:

\[\Delta w_i(t) = \chi\Delta t(-b_i(t) w_i(t) + \frac{1}{N}\sum_j b_j(t)w_j(t)) + \sum_j c_{ij}(t)\sqrt{\gamma\Delta t} \eta_{ij}(t)\min(w_i(t), w_j(t))\]

where now $b_i(t)$ represents an event of taxation over agent $i$ and $c_{ij}(t)$ represents an event of wealth exchange between agents $i$ and $j$. To see the ODE system that describes an equivalent dynamics, see the documentation of the solve_ode_net function.

Returns

w_t::Matrix{Real}: Time series of the wealth distribution. Each row is a checkpoint.

Callbacks

In the context of Monte Carlo simulations, a callback is a function that is called at the end of each iteration of the simulation. This can be used to monitor the progress of the simulation with any desired metrics. The callbacks functions take the state of the simulation as an argument, i.e. f(w) where w is the wealth distribution at the current iteration. It is easy to define custom callbacks by defining a function that takes the wealth distribution as an argument and returns something (a number, a vector, etc.). The function can then be passed to the mc_simulator function as a callback.

Some predefined callbacks are available in the YardSale module, which are commonly used in the study of wealth distributions. These are:

YardSale.get_R — Method

get_R(w::Vector{<:Real})

Calculate the number of agents with wealth above the poverty line.

Arguments

w::Vector{<:Real}: Wealth distribution.

Returns

R::Int: Number of agents with wealth above the poverty line (mean wealth).

YardSale.get_gini — Method

get_gini(w::Vector{<:Real})

Calculate the Gini coefficient of the wealth distribution.

Arguments

w::Vector{<:Real}: Wealth distribution.

Details

The Gini coefficient is a measure of statistical dispersion intended to represent the income or wealth distribution of a nation's residents, and is the most commonly used measure of inequality. It ranges from 0 (perfect equality) to 1 (perfect inequality). The formula for the Gini coefficient is:

\[g = \frac{1}{2NW} \sum_{i,j}^{N} |w_i - w_j|\]

where N is the number of agents, W is the total wealth, and w_i is the wealth of agent i.

Returns

gini::Real: Gini coefficient.

YardSale.get_lorenz — Method

get_lorenz(w::Vector{<:Real})

Calculate the Lorenz curve of the wealth distribution.

Arguments

w::Vector{<:Real}: Wealth distribution.

Details

The Lorenz curve is a graphical representation of the distribution of income or wealth. It plots the cumulative percentage of total income received against the cumulative percentage of recipients, starting with the poorest individual or household. It is used to represent income or wealth inequality. The Lorenz curve is a concave curve that lies below the diagonal line of perfect equality, which represents a situation where everyone has the same income or wealth.

Returns

lorenz::Vector{<:Real}: Lorenz curve.

YardSale.get_r — Method

get_r(w::Vector{<:Real})

Calculate the fraction of agents with wealth above the poverty line.

Arguments

w::Vector{<:Real}: Wealth distribution.

Returns

r::Real: Fraction of agents with wealth above the poverty line.

YardSale.get_u — Method

get_u(w::Vector{<:Real})

Calculate the relative mean wealth of agents with wealth above the poverty line (mean wealth)

Arguments

w::Vector{<:Real}: Wealth distribution.

Returns

u::Real: Mean wealth of agents with wealth above the poverty line.

YardSale.get_x1 — Method

get_x1(w::Vector{<:Real})

Calculate the wealth fraction of the richest agent in the system.

Arguments

w::Vector{<:Real}: Wealth distribution.

Returns

x1::Real: Wealth fraction of the richest agent.

YardSale.get_xi — Method

get_xi(w::Vector{<:Real}, i::Int)

Calculate the wealth fraction of the i-th richest agent in the system.

Arguments

w::Vector{<:Real}: Wealth distribution.
i::Int: Rank of the agent.

Returns

xi::Real: Wealth fraction of the i-th richest agent.