ODE Solvers

YardSale.dxdt_net! — Method

dxdt_net!(dxdt, x, p, t)

In-place function to calculate the derivatives of the network model.

Arguments

dxdt::Vector{<:Real}: Vector of derivatives.
x::Vector{<:Real}: Vector of wealths.
p::Tuple{<:Integer, <:Integer, Vector{Vector{<:Integer}},
Vector{<:Integer}, Vector{<:Array}, ::Real, ::Real}: Parameters.
t::Float64: Time.

Details

The network model is given by the following set of ODEs:

\[\dot{x}_i = \frac{T}{N} (- \beta_i x_i + \sum_j^N \beta_j x_j) + \sum_j^N \kappa_{ij} (x_i - x_j) \min(x_i, x_j)\]

where $x_i$ is the wealth of node $i$, $T$ is temperature, $N$ is the number of nodes, $\beta_i$ is the redistribution parameter of node $i$, $\kappa_{ij}$ is the exchange parameter between nodes $i$ and $j$.

YardSale.solve_ode_net — Method

solve_ode_net(g, tspan, integrator, interaction_mode, taxation_mode, T, seed; kwargs...)

Solves the ODE for the network model.

Arguments

g::SimpleGraph{<:Integer}: Graph.
tspan::Tuple{Float64, Float64}: Tuple with initial and final time.
interaction_mode::String: Interaction mode.
taxation_mode::String: Taxation mode.
T::Real: Temperature.
seed::Integer: Random seed.

Optional arguments

`integrator::SciMLAlgorithm`: An instante of the integrator to use. Default is Tsit5().
initial_conditions::String=nothing: Initial condition. Options are
"uniform", "random", "noisy" and "custom". Default is "uniform". If "custom" is
chosen, the x0 argument must be provided.
x0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution.
`kwargs...`: Additional arguments for the solver.

Details

This function solves the ODE for the network model. It calculates the kappa and beta parameters, sets the initial conditions, and solves the ODE. It returns the solution in the standard DifferentialEquations.jl format. A brief description of the ODE approximation model can be found in the dxdt_net! function. The function uses the DifferentialEquations.jl package to solve the ODE. All the parameters for the solver can be passed as keyword arguments.

The initial conditions can be set to "noisy", "random", "uniform", or a custom vector.

"noisy": Initial conditions are set to a random value around 1/N.

$x = (1/N) * (1 + \epsilon)$ where ϵ is white noise with μ=0 and σ=0.01.

"random": Initial conditions are set to a random value, normalized to sum to 1.

"uniform": Initial conditions are set to 1/N.

"custom": Initial conditions are set to a custom vector. The vector must sum to 1.

Also, $x_i$ must be positive for all $i$.

Returns

sol::ODESolution: Solution of the ODE.

Examples

using DifferentialEquations, Graphs, YardSale
# Create a graph
g = erdos_renyi(100, 0.1, seed=42)
interaction_mode, taxation_mode = "A","A"
T = 1.0
seed = 42
tspan = (0.0, 10.0)
sol1 = solve_ode_net(g, tspan, interaction_mode, taxation_mode, T, seed)
sol2 = solve_ode_net(g, tspan, interaction_mode, taxation_mode, T, seed; integrator=RK4())
sol3 = solve_ode_net(g, tspan, interaction_mode, taxation_mode, T, seed;
integrator=RK4(), reltol=1e-6, abstol=1e-6)

YardSale.solve_ode_net_SS — Method

solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed;
initial_conditions="noisy",integrator=Tsit5(), kwargs...)

Solves the ODE for the network model using a steady state solver.

Arguments

g::SimpleGraph{<:Integer}: Graph.
interaction_mode::String: Interaction mode.
taxation_mode::String: Taxation mode.
T::Real: Temperature.
seed::Integer: Random seed.

Optional arguments

`initial_conditions::Union{String, Vector{<:Real}}`="noisy": Initial conditions
for the ODE. Default is "noisy", which sets the initial conditions to a random value
around 1/N. Other options are "random" and "uniform". If a vector is passed, it will be
used as the initial conditions.
`x0::Union{Nothing, Vector{<:Real}}`=nothing: Initial wealth distribution.
`integrator::SciMLAlgorithm`: An instance of the integrator to use. Default is Tsit5().
`kwargs...`: Additional arguments for the solver.

Details

This function solves the ODE for the network model using a steady state solver. It is similar to solve_ode_net, but it uses the DynamicSS solver to find the steady state. Instead of an ODE problem, it uses a SteadyStateProblem to solve the ODE. Reference: https://docs.sciml.ai/DiffEqDocs/stable/types/steadystatetypes/

As in solve_ode_net, the initial conditions can be set to "noisy", "random", "uniform", or a custom vector.

Returns

sol::ODESolution: Solution of the ODE.

Examples

using DifferentialEquations, Graphs, YardSale
# Create a graph
g = erdos_renyi(100, 0.1, seed=42)
interaction_mode, taxation_mode = "A","A"
T = 1.0
seed = 42
sol1 = solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed)
sol2 = solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed; integrator=RK4())
sol3 = solve_ode_net_SS(g, interaction_mode, taxation_mode, T, seed;
                            integrator=RK4(),reltol=1e-6, abstol=1e-6
                            )