Utils

YardSale.get_kappa_beta — Method

get_kappa_beta(g::SimpleGraph{<:Integer},interaction_mode::String,taxation_mode::String)

Return the κ matrix and the β vector for the ODEs, given a graph, an Interaction Mode and a Taxation Mode.

Arguments

g::SimpleGraph{<:Integer}: Undirected graph.
interaction_mode::String: Interaction Mode. Can be "A" or "B".
taxation_mode::String: Taxation Mode. Can be "A" or "B".

Returns

Tuple{Vector{Float64},Vector{Float64}}: κ matrix and β vector.

Examples

g = complete_graph(4)
YardSale.get_kappa_beta(g,"A","A")

YardSale.EYSM_base_redistribution — Method

EYSM_base_redistribution(wi, W_N, chif_N)

Calculates the redistribution term in the EYSM model for an agent i with wealth wi.

Arguments

wi::Real: Wealth of agent i.
W_N::Real: Total wealth per agent.
chif_N::Real: Taxation and redistribution rate.

Details

In the EYSM model, the redistribution term of the wealth update of agent i $is$ given by

\[\frac{\chi f}{N}(\frac{W}{N} - w_i)\]

where f is the fraction of the minimum wealth exchanged between agents, and $f$ is playing the role of the Δt term in the Fokker-Planck approach, as the temporal unit of time of the Monte Carlo simulation. $chi$ represents the taxation and redistribution rate and N is the number of agents.

Returns

redist::Real: Redistribution term for agent i.

YardSale.mc_set_initial_conditions — Function

mc_set_initial_conditions(N, W_N, initial_conditions,w=nothing)

Set the initial conditions for the Monte Carlo simulation.

Arguments

N::Integer: Number of agents.
W_N::Real: Total wealth per agent.
initial_conditions::String: Initial condition. Options are "uniform", "random", "noisy"
and "custom".
w::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution. Only used if
initial_conditions="custom". Default is nothing.

Returns

w::Vector{Real}: Initial wealth distribution.

YardSale.Δw — Method

Δw(f, wi, wj)

Calculate the amount of wealth exchanged between agents i and j in the EYSM model.

Arguments

f::Real: Fraction of the minimum wealth exchanged between agents i and j.
wi::Real: Wealth of agent i.
wj::Real: Wealth of agent j.

Returns

Δw::Real: Amount of wealth exchanged between agents i and j.

YardSale.ηij — Method

ηij(wi, wj, zeta_W)

Calculate the stochastic variable ηij in the EYSM model.

Arguments

wi::Real: Wealth of agent i.
wj::Real: Wealth of agent j.
zeta_W::Real: Normalized bias towards the richest agent.

Details

The expected value of ηij is given by

\[\langle\eta_{ij}\rangle = \zeta \frac{w_i - w_j}{W}\]

This variable is used to determine the direction of the wealth exchange between agents i and j, and can take values ±1.

Returns

ηij::Real Stochastic variable ηij.