Utils

Additional functions for the functionalities of the package.

YardSale.get_kappa_beta — Method

get_kappa_beta(g::SimpleGraph{<:Integer},interaction_mode::String,taxation_mode::String)

Return the κ matrix and the β vector for the ODEs, given a graph, an Interaction Mode and a Taxation Mode.

Arguments

g::SimpleGraph{<:Integer}: Undirected graph.
interaction_mode::String: Interaction Mode. Can be "A" or "B".
taxation_mode::String: Taxation Mode. Can be "A" or "B".

Returns

Tuple{Vector{Float64},Vector{Float64}}: κ matrix and β vector.

Examples

g = complete_graph(4)
YardSale.get_kappa_beta(g,"A","A")

YardSale.EYSM_base_redistribution — Method

EYSM_base_redistribution(wi, W_N, chif_N)

Calculates the redistribution term in the EYSM model for an agent i with wealth wi.

Arguments

wi::Real: Wealth of agent i.
W_N::Real: Total wealth per agent.
chif_N::Real: Taxation and redistribution rate.

Details

In the EYSM model, the redistribution term of the wealth update of agent i $is$ given by

\[\frac{\chi f}{N}(\frac{W}{N} - w_i)\]

where f is the fraction of the minimum wealth exchanged between agents, and $f$ is playing the role of the Δt term in the Fokker-Planck approach, as the temporal unit of time of the Monte Carlo simulation. $chi$ represents the taxation and redistribution rate and N is the number of agents.

Returns

redist::Real: Redistribution term for agent i.

YardSale.mc_set_initial_conditions — Function

mc_set_initial_conditions(N, W_N, initial_conditions,w=nothing)

Set the initial conditions for the Monte Carlo simulation.

Arguments

N::Integer: Number of agents.
W_N::Real: Total wealth per agent.
initial_conditions::String: Initial condition. Options are "uniform", "random", "noisy"
and "custom".
w::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution. Only used if
initial_conditions="custom". Default is nothing.

Returns

w::Vector{Real}: Initial wealth distribution.

YardSale.Δw — Method

Δw(f, wi, wj)

Calculate the amount of wealth exchanged between agents i and j in the EYSM model.

Arguments

f::Real: Fraction of the minimum wealth exchanged between agents i and j.
wi::Real: Wealth of agent i.
wj::Real: Wealth of agent j.

Returns

Δw::Real: Amount of wealth exchanged between agents i and j.

YardSale.ηij — Method

ηij(wi, wj, zeta_W)

Calculate the stochastic variable ηij in the EYSM model.

Arguments

wi::Real: Wealth of agent i.
wj::Real: Wealth of agent j.
zeta_W::Real: Normalized bias towards the richest agent.

Details

The expected value of ηij is given by

\[\langle\eta_{ij}\rangle = \zeta \frac{w_i - w_j}{W}\]

This variable is used to determine the direction of the wealth exchange between agents i and j, and can take values ±1.

Returns

ηij::Real Stochastic variable ηij.

YardSale.ode_set_initial_conditions — Function

ode_set_initial_conditions(N, initial_conditions,x0=nothing)

Set the initial conditions for the ODE simulation.

Arguments

N::Integer: Number of agents.
initial_conditions::String: Initial condition. Options are "uniform", "random", "noisy"
and "custom".
x0::Union{Nothing, Vector{<:Real}}=nothing: Initial wealth distribution. Only used if
initial_conditions="custom". Default is nothing.

Returns

x::Vector{Real}: Initial wealth distribution.

YardSale.get_giant_component — Method

get_giant_component(g::SimpleGraph)

Returns the giant component of a graph.

YardSale.get_graph_data — Method

get_graph_data(g::SimpleGraph)

Returns the number of nodes, the number of links, and the edgelist of the giant component of a graph.

Details

If the graph has more than one connected component, a warning is issued, and the giant component is used for the simulation.

YardSale.x1_meanField — Method

x1_meanField(T::Real,N::Int)

Mean field prediction for the wealth of the richest agent in the system as a function of the temperature T. This is the finite size case, i.e. N is finite.

Arguments

`T::Real`: temperature of the system
`N::Int`: number of agents in the system

Returns

`x1::Real`: mean field prediction for the wealth of the richest agent

YardSale.x1_meanField — Method

x1_meanField(T::Real)

Mean field prediction for the wealth of the richest agent in the system as a function of the temperature T. This is the thermodinamic limit case, i.e. N -> ∞.

Arguments

`T::Real`: temperature of the system

Returns

`x1::Real`: mean field prediction for the wealth of the richest agent

Data analysis

Miscellaneous functions for data analysis. Common tools for the analysis of Monte Carlo and ODE simulations.

YardSale.get_avg_x1 — Method

get_avg_x1(x::AbstractMatrix, dims::Vararg{Integer}=2)

Get the temporal average of the richest node and the variance.

Arguments

x1::AbstractMatrix: Matrix of the MC simulation.
dims::Vararg{Integer}: Dimension(s) to get the richest node.
It can be an integer or a tuple of integers.

Returns

avg_x1::Tuple{Float64, Float64}: Tuple with the average and variance of the richest node.

YardSale.get_x1 — Method

get_x1(x::AbstractMatrix, dims::Union{Integer, Tuple{Integer, Integer}}=2)
Get the richest node at each time step.

Arguments

x::AbstractMatrix: Matrix of the MC solution.
dims::Union{Integer, Tuple{Integer, Integer}}: Dimension(s) to get the richest node.
It can be an integer or a tuple of integers.

Returns

x1::AbstractArray: Richest node at each time step.

YardSale.get_lambda — Method

get_lambda(g::SimpleGraph{<:Integer}, interaction_mode::String,
taxation_mode::String, T::Real, x_ss::Vector{<:Real})

Get the Lambda matrix for the ODE system.

Arguments

g::SimpleGraph{<:Integer}: Graph object.
interaction_mode::String: Interaction mode.
taxation_mode::String: Taxation mode.
T::Real: Taxation rate.
x_ss::Vector{<:Real}: Steady state vector.

Returns

Lambda::Matrix{Float64}: Lambda matrix.

Details

Coming soon.

YardSale.get_max_eigenvalue — Method

get_max_eigenvalue(Lambda::Array{<:Real})

Get the maximum real part of the eigenvalues of the a real matrix. We use this function to determine the stability of the ODE system.

Arguments

Lambda::Array{<:Real}: Matrix.

Returns

max_eigenvalue::Real: Maximum real part of the eigenvalues.

YardSale.rescale_t — Method

rescale_t(t::AbstractVector,N::Integer)

Rescale the ODE time vector to MC steps.

Arguments

t::AbstractVector: Time vector of the ODE solution.
N::Integer: Number of nodes in the network.

Returns

rescaled_t::AbstractVector: Rescaled time vector.